fix: remediate 57 confirmed Low-severity findings
Phase 4 of the review-and-hardening pass recorded in
docs/Arcanum.Review.20260810.md, completing the plan across 34 subsystems.
Each fix was made by TDD in an isolated worktree.

Two are larger than their Low severity suggests.

Fetched web content was concatenated straight into the research synthesis
prompt with no untrusted-data fencing, so a hostile page could address the
synthesising model directly. Every search answer and page body now goes through
SystemPromptBuilder.AppendUntrusted, and the attacker-controlled URL and title
moved inside the fence too -- only the code-owned "Source [n]" label stays
outside, because citation mapping needs it. The trailing instruction is
reserved out of the budget before any source is appended, so an oversized page
can no longer truncate the instruction away, and a block that will not fit once
framed is re-framed rather than dropped.

ProgressiveContextMaintainer was a stub that reported fabricated results: it
returned MessagesRemoved=2 while removing nothing, and neither injected
dependency was used. Deleted, along with its git-tracked zero-byte test file.
The live path (TurnContextGuards.TryTrimOldestToolExchanges) already covers the
behaviour and is tested. No regression test survives this one, because the
demonstration test was deleted with the type it exercised.

Also: a synthetic ask_human denial published the denial text in the tool call's
argumentsJson field; output-stage guardrail rejections reported "Input
rejected"; LexiconEntityExtractor logged the raw model response unclipped on a
parse failure; and several dead types that were shadowed by live private nested
ones were removed.

CI now compiles TheForge.Ux. Its test suite stays excluded while it is
repaired, but both release workflows package and ship the project, so without
compiling it a break merges green. DESIGN 13.8 updated to match.

Three merge notes, since the same defect was often found at more than one
severity and fixed independently from the same base.

Where both sides added a distinct new test, both were kept -- but naive
concatenation does not respect C# block structure. It detached one Codex test's
body onto the next test's signature and spliced a new rate-limiter test inside
a nested helper class, swallowing that class's closing brace. Both were rebuilt
by hand; tracing brace depth per method is what localises this.

The toolError public-error text now reads Data ?? Message rather than Data
alone, keeping the fallback for producers that only populate Message.

Arcanum.Tests 7295 passed / 34 skipped, Compendium.Tests 151, TheForge.Tests
543, full solution build clean.

+++ src/RetroDownfall.Compendium.Ux/ViewModels/GenericSettingFieldViewModel.cs
@@ -146,7 +146,9 @@ public sealed partial class GenericSettingFieldViewModel : ObservableObject
         {
             SettingKind.Int => (int)Math.Round(value),
             SettingKind.Long => (long)Math.Round(value),
-            SettingKind.Float => (float)value,
+            // Float settings are money backed by decimal properties. Narrowing to float would keep only
+            // ~7 significant digits, silently rounding a six-figure limit typed to the cent.
+            SettingKind.Float => (decimal)value,
             _ => value,
         };
